gpt_multiagent_python_openai_jira_social_plus/requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
openai==1.40.6

apscheduler==3.10.4
scikit-learn==1.5.1
